Adding extra features and info to MovieDetailsFragment.kt

+++ app/src/main/java/com/daveloper/moviesapp/data/model/entity/ReviewAuthor.kt
@@ -6,5 +6,5 @@ data class ReviewAuthor (
     @SerializedName("name") var name: String? = "",
     @SerializedName("username") var userName: String? = "",
     @SerializedName("avatar_path") var userImg: String? = "",
-    @SerializedName("rating") var rating: Int? = -1,
+    @SerializedName("rating") var rating: Float? = -1.0f,
 )
